move the builder in a sub crate
this way we'll be able to make custom builder releases.

Since there is a circular dependency on the builder and navitia_model,
to use the builder in navitia_model's test, we include directly the
builder's sources via #[path]
