+++ src/lib.rs
@@ -49,7 +49,7 @@ pub mod test_utils;
 pub mod transfers;
 
 #[cfg(test)]
-#[path = "../model-builder/src/lib.rs"]
+#[path = "../model-builder/src/builder.rs"]
 pub mod model_builder;
 
 /// The error type used by the crate.
